requirements.txt
Django>=5.0,<6.0
gunicorn
python-dotenv
django-import-export
django-livereload-server
